+++ src/helpers/strOps.js
@@ -1,0 +1,1 @@
+export const clearWhiteSpace = ({str}) => str.replace(/\s+/g, '')